fixed a few routes

+++ src/app/api/deletewill/route.ts
@@ -4,26 +4,35 @@ import WillModal from "@/models/will";
 export async function DELETE(request: Request) {
     await connectToDatabase();
 
-    try {
-        const body = await request.json();
-        const { sender, reciver } = body;
+    const { searchParams } = new URL(request.url);
+    const reciver = searchParams.get('receiver');
+    const sender = searchParams.get('sender');
 
-        const result = await WillModal.findOneAndDelete({ sender, reciver });
+    if (sender && reciver) {
+        try {
 
-        if (!result) {
-            return new Response(JSON.stringify({
-                success: false,
-                message: "No will found for the given sender and receiver."
-            }), {
-                status: 404,
+
+            const result = await WillModal.findOneAndDelete({ sender, reciver });
+
+            if (!result) {
+                return new Response(JSON.stringify({
+                    success: false,
+                    message: "No will found for the given sender and receiver."
+                }), {
+                    status: 404,
+                });
+            }
+
+            return new Response(JSON.stringify({ success: true, message: "Will deleted successfully." }), {
+                status: 200,
+            });
+        } catch (error) {
+            return new Response(JSON.stringify({ success: false, message: "Failed to delete will" }), {
+                status: 500,
             });
         }
-
-        return new Response(JSON.stringify({ success: true, message: "Will deleted successfully." }), {
-            status: 200,
-        });
-    } catch (error) {
-        return new Response(JSON.stringify({ success: false, message: "Failed to delete will" }), {
+    } else {
+        return new Response(JSON.stringify({ success: false, message: "failed to parse parameeters" }), {
             status: 500,
         });
     }
